refactor: rename frontend pages to views and disable native streamlit multipage sidebar
