+++ src/brad/frontend/views/balance_entry.py
