+++ src/brad/frontend/views/entity_management.py
