+++ src/brad/frontend/views/valuation_entry.py
